# api-docs.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "FAM Backend API",
    "version": "1.0.0",
    "description": "Family event management platform — three-role auth (user/admin/organizer), family tree, QR join, notifications."
  },
  "servers": [
    {
      "url": "http://localhost:8000"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "RelationshipObject": {
        "type": "object",
        "properties": {
          "_id": { "type": "string" },
          "eventId": { "type": "string" },
          "person1": { "type": "string" },
          "person2": { "type": "string" },
          "relationType": { "type": "string", "example": "parent" },
          "familySide": { "type": "string", "enum": ["groom", "bride", "common"] },
          "isValidated": { "type": "boolean" },
          "marriageDate": { "type": "string", "format": "date-time", "nullable": true }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/organizer/login": {
      "post": {
        "summary": "Organizer login with email + password",
        "description": "Rate limited: 5 req/15 min. Returns 24h access token.",
        "tags": ["Organizer"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email", "password"],
                "properties": {
                  "email": { "type": "string", "example": "org@example.com" },
                  "password": { "type": "string", "example": "tempPass123" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Returns access token (24h) and organizer object",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "token": { "type": "string" },
                    "organizer": {
                      "type": "object",
                      "properties": {
                        "id": { "type": "string" },
                        "name": { "type": "string" },
                        "email": { "type": "string" },
                        "role": { "type": "string", "example": "organizer" },
                        "assignedEvent": { "type": "string" },
                        "permissions": { "type": "array", "items": { "type": "string" } }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": { "description": "Invalid credentials" }
        }
      }
    },
    "/api/organizer/logout": {
      "post": {
        "summary": "Organizer logout",
        "tags": ["Organizer"],
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Logged out",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "message": { "type": "string", "example": "Organizer logged out successfully" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/organizer/profile": {
      "get": {
        "summary": "Get organizer profile with populated assignedEvent",
        "tags": ["Organizer"],
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "Organizer profile object" }
        }
      }
    },
    "/api/organizer/assigned-event": {
      "get": {
        "summary": "Full event document with populated participants and creator",
        "tags": ["Organizer"],
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "Assigned event details" },
          "404": { "description": "No event assigned" }
        }
      }
    },
    "/api/organizer/participants": {
      "get": {
        "summary": "Participants of the assigned event (username, email, dob, profileImage, phone, location, addedPeopleCount)",
        "tags": ["Organizer"],
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "Array of participants" }
        }
      }
    },
    "/api/organizer/stats": {
      "get": {
        "summary": "Dashboard statistics for the assigned event",
        "tags": ["Organizer"],
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Event stats",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "stats": {
                      "type": "object",
                      "properties": {
                        "totalParticipants": { "type": "integer", "example": 42 },
                        "totalRelationships": { "type": "integer", "example": 80 },
                        "eventActive": { "type": "boolean" },
                        "eventDate": { "type": "string", "format": "date-time" }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/organizer/pending-approvals": {
      "get": {
        "summary": "Users waiting for approval to join the event",
        "tags": ["Organizer"],
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "Array of pending users" }
        }
      }
    },
    "/api/organizer/event/get-approval-mode": {
      "get": {
        "summary": "Get current approval mode of the assigned event",
        "tags": ["Organizer"],
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Current approval mode",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "approvalMode": { "type": "string", "enum": ["auto", "manual"] }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/organizer/event/approval-mode": {
      "patch": {
        "summary": "Update approval mode for the assigned event",
        "tags": ["Organizer"],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["approvalMode"],
                "properties": {
                  "approvalMode": { "type": "string", "enum": ["auto", "manual"] }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Approval mode updated" }
        }
      }
    },
    "/api/organizer/approve/{userId}": {
      "patch": {
        "summary": "Approve a pending user — moves from pendingApprovals to participants and sends notification",
        "tags": ["Organizer"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "in": "path", "name": "userId", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "User approved" },
          "404": { "description": "User not in pending approvals" }
        }
      }
    },
    "/api/organizer/reject/{userId}": {
      "patch": {
        "summary": "Reject a pending user — removes from pendingApprovals and sends rejection notification",
        "tags": ["Organizer"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "in": "path", "name": "userId", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "User rejected" }
        }
      }
    },
    "/api/organizer/manual-add-user": {
      "post": {
        "summary": "Create or find user by email and add them directly to event participants (bypasses approval)",
        "tags": ["Organizer"],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name", "email", "dob"],
                "properties": {
                  "name": { "type": "string", "example": "Alice Smith" },
                  "email": { "type": "string", "example": "alice@example.com" },
                  "dob": { "type": "string", "format": "date", "example": "1985-03-10" },
                  "bloodGroup": { "type": "string", "example": "B+" },
                  "profession": { "type": "string", "example": "Teacher" },
                  "location": { "type": "string", "example": "Chennai" },
                  "gender": { "type": "string", "enum": ["male", "female", "other"] },
                  "phone": { "type": "string", "example": "+911122334455" },
                  "socialMediaLink": { "type": "string", "example": "https://facebook.com/alice" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "User added to event participants" },
          "404": { "description": "No event assigned to organizer" }
        }
      }
    },
    "/api/organizer/schedule": {
      "patch": {
        "summary": "Replace the entire event schedule",
        "tags": ["Organizer"],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["schedule"],
                "properties": {
                  "schedule": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "title": { "type": "string", "example": "Welcome Reception" },
                        "description": { "type": "string" },
                        "dateTime": { "type": "string", "format": "date-time", "example": "2026-06-15T18:00:00Z" },
                        "location": { "type": "string", "example": "Grand Hall" }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Schedule updated" }
        }
      }
    },
    "/api/organizer/unvalidated-relationships": {
      "get": {
        "summary": "All relationships in the assigned event where isValidated=false",
        "tags": ["Organizer"],
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "Array of unvalidated relationships" }
        }
      }
    },
    "/api/organizer/validate-relationship/{id}": {
      "patch": {
        "summary": "Validate or reject a proposed relationship",
        "tags": ["Organizer"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "in": "path", "name": "id", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["isValidated"],
                "properties": {
                  "isValidated": { "type": "boolean", "description": "true to validate, false to invalidate" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Relationship validation status updated",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "message": { "type": "string", "example": "Relationship validated successfully" },
                    "relationship": { "$ref": "#/components/schemas/RelationshipObject" }
                  }
                }
              }
            }
          },
          "404": { "description": "Relationship not found" }
        }
      }
    },
...
  },
...
}